pyproject.toml
[project]
name = "agent-evaluator"
version = "0.1.0"
description = "基于ragas的智能体测评框架"
authors = [
    {name = "CaptainJi", email = "jiqing19861123@163.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "ragas>=0.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "rich>=13.0",
    "langchain-openai>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
agent-eval = "agent_evaluator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_evaluator"]

